Refactoring ConfigureRepos & adding FetchRepos
- Splitting out the repo fetching logic into its own class, as this will be needed to support a repo validation script that will be added in the next commit.
- Refactored ConfigureRepos to utilise FetchRepos where required.
